apply_to_2d workflow
